Add a changelog
... because we should have some sort of history.

I'm not picky about format. This one is called HISTORY.rst because I
copied it from gapipy. It can be CHANGELOG.md, really, if anyone
expresses a preference.

I like *some* place to keep track of unlreleased stuff that has landed in
master. Anyone have feelings about this being in or out of this file?