Don't validate webhook HREF values
In practice, if we reject a webhook here because of an unexpected href
value it will simply we redelivered to us over and over again and we'll
reject it over and over again until some timeout is reached.

(I believe at time of writing, we will attempt to deliver events for
three days before giving up)

If API backends are sending out bad HREFs we should correct it at the
source, having our webhook handlers reject these events is not a
solution.

Interestingly this came up because we had some manually crafted webhooks
that were sent out in order to correct a problem. There was an error in
the HREF so they've been bouncing around getting rejected and retried
for the past little while. (We sent the correct webhooks after the
fact.)