Test on all the Django 4.x versions (and Python 3.11)
I'm rearranging the tox.ini envlist. I found it much easier to think
about one Django at a time and then just go read the PyPI metadata to
understand what Python versions are supported by the latest relase of
each.

We had been testing the combination of Django 1.11 and Python 3.8, but
actually that's not supported (I just left it in for now, but could see
an argument for or against):
    https://docs.djangoproject.com/en/1.11/releases/1.11/#python-compatibility
    https://pypi.org/project/Django/1.11.29/